tests/wpt/web-platform-tests/offscreen-canvas/fill-and-stroke-styles/2d.pattern.repeat.null.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.pattern.repeat.null</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>

<h1>2d.pattern.repeat.null</h1>
<p class="desc"></p>


<script>
var t = async_test("");
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

_assert(ctx.createPattern(offscreenCanvas, null) != null, "ctx.createPattern(offscreenCanvas, null) != null");

t.done();

});
</script>
